LayoutTests/imported/w3c/web-platform-tests/html/semantics/interactive-elements/the-dialog-element/dialog-cancel-with-input-expected.txt

CONSOLE MESSAGE: ReferenceError: Can't find variable: action
Test dialog modal is closed by escape key with input focused

Hello World


FAIL Test dialog modal is closed by escape key with input focused ReferenceError: Can't find variable: action
